Event nav bar menu replaced wih sidebar, displayed in all event related views
Event dropdown list now lists events moderated for moderators

Renamed 'Eventos abertos' to 'Eventos em andamento' and added call to status update method

Added admin role checker for event deletion buttons

Implemented isMod() method in Event

Reimplemented event factory

Allowed moderators to cancel subscriptions

Reworked logic checking in subscription and submission buttons

Removed event's title from moderator checkboxes and added help text

Added logic to the  event observer to avoid deleting the placeholder logo

+++ app/Observers/EventObserver.php
@@ -20,9 +20,14 @@ class EventObserver
      */
     public function updated(Event $event): void
     {
-        //dd($event->logo);
         if ($event->isDirty('logo')) {
-            Storage::disk('public')->delete($event->getOriginal('logo'));
+            $originalLogo = $event->getOriginal('logo');
+            $placeholderLogo = 'event_logos/Placeholder.jpg'; // Change this to match your placeholder file name
+    
+            // Check if the logo is not the placeholder before deleting
+            if ($originalLogo !== $placeholderLogo) {
+                Storage::disk('public')->delete($originalLogo);
+            }
         }
     }
 
@@ -31,8 +36,13 @@ class EventObserver
      */
     public function deleted(Event $event): void
     {
-        if (! is_null($event->logo)) {
-            Storage::disk('public')->delete($event->logo);
+        if (!is_null($event->logo)) {
+            $placeholderLogo = 'event_logos/Placeholder.jpg'; // Change this to match your placeholder file name
+    
+            // Check if the logo is not the placeholder before deleting
+            if ($event->logo !== $placeholderLogo) {
+                Storage::disk('public')->delete($event->logo);
+            }
         }
     }
